refactor: member module package cleanup
