+++ family/src/main/java/com/oing/repository/FamilyRepository.java
@@ -1,6 +1,6 @@
 package com.oing.repository;
 
-import com.oing.domain.model.Family;
+import com.oing.domain.Family;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FamilyRepository extends JpaRepository<Family, String> {
