+++ member/src/main/java/com/oing/controller/AuthController.java
@@ -3,7 +3,7 @@ package com.oing.controller;
 import com.oing.domain.*;
 import com.oing.exception.DomainException;
 import com.oing.exception.ErrorCode;
-import com.oing.domain.model.Member;
+import com.oing.domain.Member;
 import com.oing.dto.request.CreateNewMemberRequest;
 import com.oing.dto.request.NativeSocialLoginRequest;
 import com.oing.dto.request.RefreshAccessTokenRequest;
