+++ member/src/main/java/com/oing/controller/MemberController.java
@@ -1,7 +1,7 @@
 package com.oing.controller;
 
 import com.oing.domain.PaginationDTO;
-import com.oing.domain.model.Member;
+import com.oing.domain.Member;
 import com.oing.dto.request.PreSignedUrlRequest;
 import com.oing.dto.request.UpdateMemberNameRequest;
 import com.oing.dto.request.UpdateMemberProfileImageUrlRequest;
