+++ member/src/main/java/com/oing/domain/Member.java
@@ -1,4 +1,4 @@
-package com.oing.domain.model;
+package com.oing.domain;
 
 import com.oing.domain.DeletableBaseAuditEntity;
 import jakarta.persistence.Column;
