+++ member/src/main/java/com/oing/domain/key/MemberDeviceKey.java
@@ -1,6 +1,5 @@
-package com.oing.domain.model.key;
+package com.oing.domain.key;
 
-import com.oing.domain.SocialLoginProvider;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -10,13 +9,13 @@ import java.io.Serializable;
 /**
  * no5ing-server
  * User: CChuYong
- * Date: 2023/11/27
+ * Date: 2024/01/02
  * Time: 11:43 AM
  */
 @AllArgsConstructor
 @NoArgsConstructor
 @EqualsAndHashCode
-public class SocialMemberKey implements Serializable {
-    private SocialLoginProvider provider;
-    private String identifier;
+public class MemberDeviceKey implements Serializable {
+    private String memberId;
+    private String fcmToken;
 }
