+++ member/src/main/java/com/oing/repository/MemberDeviceRepository.java
@@ -1,7 +1,7 @@
 package com.oing.repository;
 
-import com.oing.domain.model.MemberDevice;
-import com.oing.domain.model.key.MemberDeviceKey;
+import com.oing.domain.MemberDevice;
+import com.oing.domain.key.MemberDeviceKey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberDeviceRepository extends JpaRepository<MemberDevice, MemberDeviceKey> {
