+++ member/src/main/java/com/oing/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.oing.repository;
 
-import com.oing.domain.model.Member;
+import com.oing.domain.Member;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
