+++ member/src/main/java/com/oing/service/MemberDeviceService.java
@@ -1,7 +1,7 @@
 package com.oing.service;
 
-import com.oing.domain.model.MemberDevice;
-import com.oing.domain.model.key.MemberDeviceKey;
+import com.oing.domain.MemberDevice;
+import com.oing.domain.key.MemberDeviceKey;
 import com.oing.repository.MemberDeviceRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
