+++ member/src/test/java/com/oing/domain/model/SocialMemberTest.java
@@ -1,6 +1,8 @@
 package com.oing.domain.model;
 
+import com.oing.domain.Member;
 import com.oing.domain.SocialLoginProvider;
+import com.oing.domain.SocialMember;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
